changing to tsx files
