+++ AppRegistration.tsx
